+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SDataTable` component now supports `pageSizes` prop to set the page sizes.
 
 ## 2.0.0-beta.37 - 2023-11-14
 ### Added