+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SDataTable` component now supports `pagination` prop to set the pagination.
 - `SDataTable` component now supports `pageSizes` prop to set the page sizes.
 
+### Changed
+- `SSidebar` styles updated.
+
 ## 2.0.0-beta.37 - 2023-11-14
 ### Added
 - `SCombobox` component now supports `flipOptions` prop to flip the options list.